Finish password_checker.py
This password_checker uses a for loop to iterate through the password
inputed as a list and then checks each character to see if it's an uppercase,
lowercase, or special character. It also checks the length of the password
If the password is missing any of those characters or it is less than 9
characters in length, the program results in an error telling the user
it's password is weak.